+++ app/src/main/java/com/ykis/ykispam/ui/screens/service/list/TotalServiceItem.kt
@@ -9,5 +9,5 @@ data class TotalServiceDebt(
     val contentDetail: ContentDetail,
     val icon : ImageVector,
     val color : Color,
-    val debt : Float
+    val debt : Double
 )
